Do not show all payments to every client
This was a big bug where every user of the system saw every payment anyone ever made.  This was because current_user.customer_id returned nil, which ended up returning all payments with a nil customer_id.

We protect against this in payment.rb by adding a where.not for customer_id being nil.