Refactored form
There were a number of issues I noted when form was originally created.
Although it had nothing to do with suggested hours, I still thought it
was good to change a few things to make it simple. Most necessary
info could be gained simply by accessing the engagement. Switched form
to reflect that simplicity.