Applied changes to forms so they view correctly
Chagnes were made so that invoice form for directors show corretly.
Also, minor changes have been made so that forms show nicely on the
dashboard. Most important to note is that the partial "invoice_form" now
takes a local variable argument to see if it should render as center.
Changes to other forms were made so the edges of the forms line up
correctly.