Created suggestion form for dashboard
A suggestion form that appears if a tutor has engagements with low balances has been added. To achieve this, the proper route to resources was added to the routes. A suggestion controller with a create action was created. Necessary logic was added to the dashboard controller. Finally, all the html that was affected was altered to make everything work.