Moved favicon to asset pipeline
Favicon was in the pulic folder which is a technique that will be
deprecated by Rails going forward. The new method is to place the
favicon in the app/assets/images folder. The `favicon_link_tag` will
automatically find the favicon and serve. References in the view files to
the favicon was also changed to reflect changes.