Necessary Migrations for Payment Feature
Some migrations were necessary as I retooled the payment feature. The unsused status column for invoices was changed to be a string while tutor pay was added. Since the amount for payments was not being kept in the database as cents, this was also changed. I added the Payment#amount so that features that depends on this usage are still able to access the necessary info.