Replaced hardcoded, environement specific Id's
Certain Id's in testing factories were hardcoded. This is a problem
since those id's are specific to the developer's environment. If they have a
different sandbox environment for Dwolla, writing a new test will fail
since it will not reach the correct API endpoints. Therefore, the
harcoded lines were changed to ENV variables and placed in the
`.env.test` file which was also added to the `.gitignore` file.