Updated Payment Service and Controller
The payment service and controllers were not handling one time payments
correctly because of how it retrieved the customer. Some refactoring was
done to make the code understandable. Removed unnecessary code. Views
were changed so that payment form will not look squished when it is on
the page by itself.