Empty engagement list is ignored on Client Dash
The logic in the client dashboard view does not account for the
possibility that the client engagement list might be empty. This commit
adds a check for that as well as combine the student engagement and
client engagement for the engagement list. This is because all students
are currently clients in the database and the logic in the dash
controller does not work.