Validates presence of data for Invoice
Although our forms have validations, this can be bypassed by editing the
html. To ensure that we maintain the integrity of our data, I've added
validations to Invoices. This can also be bypassed if we skip
validations, but hopefully, if we ever do that, it will be intentional.
All tests have been updated to account for the change in validation.