Shows unassigned users on Admin Roles page
Currently, if a user has no roles, they are not shown anywhere on the
site. This makes it impossible to fix from the Admin's view. Therefore,
an unassigned category was made and Admin now has the ability to add
roles to these users. Since these users can be clients, an extra 'Add as
Client' was added.