Changes display of time into words
A private method that converts the total suggested minutes into a time
string has been added in the payments controller. The method will output
"x hour(s) (y minutes)" as a string. Hours are pluralized depending on
number of hours and minutes will be added only if the time is not an
even hour.