Refactors client dashboard with widgets
The dashboard for our clients did not contain useful information. The
dashboard was refactored to show a list of the client's tutors as well
as their balance. The balance will only show the type of balance the
user currently is engaged in. It also removes the balance in the left
menu if the user has more than two types of engagements. This is because
previously, it was displaying the sum of academic and test prep
balances. This is confusing and possibly dangerous for our business. For
clarity, this balance is only shown when the client has only one kind of
engagement.