Refactors routing of Dashboards
`get "/dashboard"` was specified globally. However, each role had a
route for dashboards also assigned. This dashboard assignment sent the
route to a specific role based method in the dashboard controller. Since
this automatically renders the appropriate file, we can get rid of the
`render :view` line as well as separate each dashboard into it's one
method. This removes complexity at the controller level. Also, the
students dashboard was never specified and therefore, the app would
break for anyone with a student role. For now, since the student is
redirected to the profile page, since there is no feature yet for
students.