Simplifies payment options for Clients
Clients with only one type of engagement was still shown options to
purchase two types of credits. This is confusing. Therefore, this pr
makes use of the `user#academic_types_engaged` to simplify the payment
options. If the user only has one type of engagement, the payment form
will lock the rates for that engagement. Else, they will be show both
options. This also removes the unnecesary complexity in the from by
removing money to currency service and simply using the monetize gem
options to convert rates to string.