Updates Payment, Invoice, Contract relations
In order to have our relations work seamlessly with the monetize gem, payments relations columns name was changed from amount_in_cents to amount_cents. Contract had an unnecessary hourly_rate column that was dropped. Invoice had the most massive change. It required the change of hourly_rate into cents as well as data changes for existing data. This was because the hourly rate was kept as dollars instead of cents. Amount also had to converted since it was formerly being recorded as a decimal instead of integers.