Implements monetize on payment related features
In keeping with the monetize gem, all payment related feature was changed to reflect the new syntax. This reduces the use of floats to 0 with regards to money. Intuitive method calls are now normative instead of complicated ones ie(`amount: amount` and not `amount_in_cents: (amount.to_f * 100).to_i`).