Changes payments to update balances/status
The single payment feature for invoices and timesheets were not
being handled correctly. Statuses were not being set to paid and the
amount being paid was not being sent. This caused a bug in our payments
feature. Balances were also not being correctly set. Updates to the view
and services cleaned up this bug.