Changes Database cleaning strategy
Formerly, a separate database cleaner strategy file was used. This was
brought into the rails helper file. Also, the strategy was changed to
load the seeds.rb file rather than have another file generate the
necessary data for the app to run. This also required moving the
subjects in the seeds.rb to the dev:seeds rake command. Also made sure
the that dev:seeds call in seeds.rb only runs in development or staging
environment.