Separate one time payment from invoice payments
Before this pr, the one time payment was being hndled in the same
controller as a payment to an invoice. This made the controller more
complex than it needed to be. Therefore, a new controller was created to
handle miscellaneous payments made that does not affect a tutor or
contractor's balance.