Sends emails async and changes method names
Instead of `deliver_now`, mails are now set to `deliver_later` which uses the sidekiq mailer queue. Certain mailer actions were named to generally to understand what was happening. Since the mailer name was descriptive of the content, the action was changed to whom the mails are sent. Previews and views were changed accordingly.