Associates ClientAccounts to Engagements
User is not bound to foreign_key client id anymore. It is accessed through the client account. One of the functionality that is lost is calling engagements directly on the user. However, by being strict about calling the account before calling engagements, we ensure every developer is clear on what is happening.