Adds specs for new features
While writing these tests, certain aspects of the apps were changed.
Notably, added a no reply email to the new client mailer since a "to"
parameter is required. Also made subject a requirement for engagements
while ensuring a student account for the Client is made if the Client
signed up as a student.