Adds payer_name to payments
Payment is not always going to have a payer and therefore, migth cause issues in view where we are showing the payer name. A method was added that tests if theres is a payer. If there is, the user's name is provided. If not, the cardholder name is provided instead.