Changes payment/refund amount calculation
For consistency, all payment calculation from rate * hours is set to
floor instead of round. This causes an issue with refunds since if
multiple refunds are being made, the possiblity of incomplete refunds
exists. Therefore, calculating the amount of refunds has been changed to
take into account whether refund is for the entire refund or not. If it
is for the entire payment, then instead of calculating it from rate *
hours, it simply takes the remaining payment amount.