Payment view accounts for old data
Some older payments are missing certain attributes because they were made before features were made. This causes the app to break since it is expecting those fields to generate the modals. This commit adds checks for those attributes so pages will not break. This was mitigated in the past because older payments were usually in the later paginations, therefore generally not rendered.